Moved FieldListModal
